+++ CHANGES.md
@@ -4,6 +4,10 @@ Changelog
 Unreleased
 ----------
 
+### Fixed
+
+- Fix bug where mixing schema directives with `@deprecated` would break.
+
 [0.6.0](https://github.com/lirsacc/py-gql/releases/tag/0.6.0) - 2020-04-01
 --------------------------------------------------------------------------